Separated QueryGraph creation and printing
