+++ src/LocalMakefile
@@ -1,7 +1,7 @@
 include src/operator/LocalMakefile
 include src/cts/LocalMakefile
 
-obj_rts:=bin/Database.o bin/Register.o bin/Table.o bin/Attribute.o $(obj_operator)
+obj_rts:=bin/Database.o bin/Register.o bin/Table.o bin/Attribute.o bin/QueryGraph.o $(obj_operator)
 
 bin/admin$(EXEEXT): bin/Admin.o $(obj_rts)
 	$(BUILDEXE)
@@ -12,5 +12,5 @@ bin/isql$(EXEEXT): bin/isql.o $(obj_rts) $(obj_sqlparser) $(obj_semana)
 bin/runsql$(EXEEXT): bin/runsql.o $(obj_rts) $(obj_sqlparser) $(obj_semana)
 	$(BUILDEXE)
 
-bin/QueryGraph$(EXEEXT): bin/QueryGraph.o $(obj_rts) $(obj_sqlparser) $(obj_semana)
+bin/QueryGraph$(EXEEXT): bin/QueryGraphTest.o $(obj_rts) $(obj_sqlparser) $(obj_semana)
 	$(BUILDEXE)
